+++ regression/scripts/concorde/concorde.py
@@ -33,6 +33,11 @@ from SUAVE.Core import (
 Data, Container,
 )
 
+import sys
+sys.path.append('../Vehicles')
+from Concorde import vehicle_setup, configs_setup
+
+
 # This is a sizing function to fill turbojet parameters
 from SUAVE.Methods.Propulsion.turbojet_sizing import turbojet_sizing
 
@@ -158,391 +163,6 @@ def base_analysis(vehicle):
     
     # done!
     return analyses    
-
-# ----------------------------------------------------------------------
-#   Define the Vehicle
-# ----------------------------------------------------------------------
-
-def vehicle_setup(source_ratio=1.):
-
-    # ------------------------------------------------------------------
-    #   Initialize the Vehicle
-    # ------------------------------------------------------------------    
-    
-    vehicle = SUAVE.Vehicle()
-    vehicle.tag = 'Concorde'    
-    
-    
-    # ------------------------------------------------------------------
-    #   Vehicle-level Properties
-    # ------------------------------------------------------------------    
-
-    # mass properties
-    vehicle.mass_properties.max_takeoff               = 185000.   # kg
-    vehicle.mass_properties.operating_empty           = 78700.   # kg
-    vehicle.mass_properties.takeoff                   = 185000.   # kg
-    vehicle.mass_properties.cargo                     = 1000.  * Units.kilogram   
-        
-    # envelope properties
-    vehicle.envelope.ultimate_load = 3.5
-    vehicle.envelope.limit_load    = 1.5
-
-    # basic parameters
-    vehicle.reference_area         = 358.25      
-    vehicle.passengers             = 100
-    vehicle.systems.control        = "fully powered" 
-    vehicle.systems.accessories    = "long range"
-    
-    
-    # ------------------------------------------------------------------        
-    #   Main Wing
-    # ------------------------------------------------------------------        
-    
-    wing = SUAVE.Components.Wings.Wing()
-    wing.tag = 'main_wing'
-    
-    wing.aspect_ratio            = 1.83
-    wing.sweeps.quarter_chord     = 59.5 * Units.deg
-    wing.thickness_to_chord      = 0.03
-    wing.taper                   = 0.
-    wing.span_efficiency         = 0.74
-    
-    wing.spans.projected         = 25.6    
-    
-    wing.chords.root             = 33.8
-    wing.total_length            = 33.8
-    wing.chords.tip              = 1.1
-    wing.chords.mean_aerodynamic = 18.4
-    
-    wing.areas.reference         = 358.25 
-    
-    wing.twists.root             = 0.0 * Units.degrees
-    wing.twists.tip              = 0.0 * Units.degrees
-    
-    wing.origin                  = [14,0,-.8]
-    wing.aerodynamic_center      = [35,0,0] 
-    
-    wing.vertical                = False
-    wing.symmetric               = True
-    wing.high_lift               = True
-    wing.vortex_lift             = True
-    wing.high_mach               = True
-    
-    wing.dynamic_pressure_ratio  = 1.0
-    
-    # add to vehicle
-    vehicle.append_component(wing)
-    
-    
-    # ------------------------------------------------------------------
-    #   Vertical Stabilizer
-    # ------------------------------------------------------------------
-    
-    wing = SUAVE.Components.Wings.Wing()
-    wing.tag = 'vertical_stabilizer'    
-    
-    wing.aspect_ratio            = 0.74      #
-    wing.sweeps.quarter_chord    = 60 * Units.deg
-    wing.thickness_to_chord      = 0.04
-    wing.taper                   = 0.14
-    wing.span_efficiency         = 0.9
-    
-    wing.spans.projected         = 6.0      #    
-
-    wing.chords.root             = 14.5
-    wing.total_length            = 14.5
-    wing.chords.tip              = 2.7
-    wing.chords.mean_aerodynamic = 8.66
-    
-    wing.areas.reference         = 33.91    #
-    
-    wing.twists.root             = 0.0 * Units.degrees
-    wing.twists.tip              = 0.0 * Units.degrees  
-    
-    wing.origin                  = [42.,0,1.]
-    wing.aerodynamic_center      = [50,0,0]    
-    
-    wing.vertical                = True 
-    wing.symmetric               = False
-    wing.t_tail                  = False
-    wing.high_mach               = True     
-    
-    wing.dynamic_pressure_ratio  = 1.0
-    
-    # add to vehicle
-    vehicle.append_component(wing)    
-
-
-    # ------------------------------------------------------------------
-    #  Fuselage
-    # ------------------------------------------------------------------
-    
-    fuselage = SUAVE.Components.Fuselages.Fuselage()
-    fuselage.tag = 'fuselage'
-    
-    fuselage.seats_abreast         = 4
-    fuselage.seat_pitch            = 1
-    
-    fuselage.fineness.nose         = 4.3
-    fuselage.fineness.tail         = 6.4
-    
-    fuselage.lengths.total         = 61.66  
-    
-    fuselage.width                 = 2.88
-    
-    fuselage.heights.maximum       = 3.32    #
-    
-    fuselage.heights.maximum       = 3.32    #
-    fuselage.heights.at_quarter_length              = 3.32    #
-    fuselage.heights.at_wing_root_quarter_chord     = 3.32    #
-    fuselage.heights.at_three_quarters_length       = 3.32    #
-
-    fuselage.areas.wetted          = 523.
-    fuselage.areas.front_projected = 7.55
-    
-    
-    fuselage.effective_diameter    = 3.1
-    
-    fuselage.differential_pressure = 7.4e4 * Units.pascal    # Maximum differential pressure   
-    
-    # add to vehicle
-    vehicle.append_component(fuselage)
-    
-        
-    # ------------------------------------------------------------------
-    #   Turbojet Network
-    # ------------------------------------------------------------------    
-    
-    # instantiate the gas turbine network
-    turbojet = SUAVE.Components.Energy.Networks.Turbojet_Super()
-    turbojet.tag = 'turbojet'
-    
-    # setup
-    turbojet.number_of_engines = 4.0
-    turbojet.engine_length     = 12.5
-    turbojet.nacelle_diameter  = 1.60
-    turbojet.areas             = Data()
-    turbojet.areas.wetted      = 12.5*1.6*8. # essentially rectangles attached to the wings
-    
-    # working fluid
-    turbojet.working_fluid = SUAVE.Attributes.Gases.Air()
-    
-    
-    # ------------------------------------------------------------------
-    #   Component 1 - Ram
-    
-    # to convert freestream static to stagnation quantities
-    
-    # instantiate
-    ram = SUAVE.Components.Energy.Converters.Ram()
-    ram.tag = 'ram'
-    
-    # add to the network
-    turbojet.append(ram)
-
-
-    # ------------------------------------------------------------------
-    #  Component 2 - Inlet Nozzle
-    
-    # instantiate
-    inlet_nozzle = SUAVE.Components.Energy.Converters.Compression_Nozzle()
-    inlet_nozzle.tag = 'inlet_nozzle'
-    
-    # setup
-    inlet_nozzle.polytropic_efficiency = 0.98
-    inlet_nozzle.pressure_ratio        = 1.0
-    
-    # add to network
-    turbojet.append(inlet_nozzle)
-    
-    
-    # ------------------------------------------------------------------
-    #  Component 3 - Low Pressure Compressor
-    
-    # instantiate 
-    compressor = SUAVE.Components.Energy.Converters.Compressor()    
-    compressor.tag = 'low_pressure_compressor'
-
-    # setup
-    compressor.polytropic_efficiency = 0.91
-    compressor.pressure_ratio        = 3.1    
-    
-    # add to network
-    turbojet.append(compressor)
-
-    
-    # ------------------------------------------------------------------
-    #  Component 4 - High Pressure Compressor
-    
-    # instantiate
-    compressor = SUAVE.Components.Energy.Converters.Compressor()    
-    compressor.tag = 'high_pressure_compressor'
-    
-    # setup
-    compressor.polytropic_efficiency = 0.91
-    compressor.pressure_ratio        = 5.0  
-    
-    # add to network
-    turbojet.append(compressor)
-
-
-    # ------------------------------------------------------------------
-    #  Component 5 - Low Pressure Turbine
-    
-    # instantiate
-    turbine = SUAVE.Components.Energy.Converters.Turbine()   
-    turbine.tag='low_pressure_turbine'
-    
-    # setup
-    turbine.mechanical_efficiency = 0.99
-    turbine.polytropic_efficiency = 0.93     
-    
-    # add to network
-    turbojet.append(turbine)
-    
-      
-    # ------------------------------------------------------------------
-    #  Component 6 - High Pressure Turbine
-    
-    # instantiate
-    turbine = SUAVE.Components.Energy.Converters.Turbine()   
-    turbine.tag='high_pressure_turbine'
-
-    # setup
-    turbine.mechanical_efficiency = 0.99
-    turbine.polytropic_efficiency = 0.93     
-    
-    # add to network
-    turbojet.append(turbine)
-      
-    
-    # ------------------------------------------------------------------
-    #  Component 7 - Combustor
-    
-    # instantiate    
-    combustor = SUAVE.Components.Energy.Converters.Combustor()   
-    combustor.tag = 'combustor'
-    
-    # setup
-    combustor.efficiency                = 0.99
-    combustor.alphac                    = 1.0     
-    combustor.turbine_inlet_temperature = 1450.
-    combustor.pressure_ratio            = 1.0
-    combustor.fuel_data                 = SUAVE.Attributes.Propellants.Jet_A()    
-    
-    # add to network
-    turbojet.append(combustor)
-
-    
-    # ------------------------------------------------------------------
-    #  Component 8 - Core Nozzle
-    
-    # instantiate
-    nozzle = SUAVE.Components.Energy.Converters.Supersonic_Nozzle()   
-    nozzle.tag = 'core_nozzle'
-    
-    # setup
-    nozzle.polytropic_efficiency = 0.95
-    nozzle.pressure_ratio        = 0.99    
-    
-    # add to network
-    turbojet.append(nozzle)
-    
-    
-    # ------------------------------------------------------------------
-    #Component 9 : thrust (to compute the thrust)
-    thrust = SUAVE.Components.Energy.Processes.Thrust()       
-    thrust.tag ='compute_thrust'
- 
-    # total design thrust (includes all the engines)
-    thrust.total_design             = 4*140000. * Units.N #Newtons
- 
-    # Note: Sizing builds the propulsor. It does not actually set the size of the turbojet
-    # Design sizing conditions
-    altitude      = 0.0*Units.ft
-    mach_number   = 0.01
-    isa_deviation = 0.
-    
-    # add to network
-    turbojet.thrust = thrust
-
-    #size the turbojet
-    turbojet_sizing(turbojet,mach_number,altitude)   
-    
-    # add  gas turbine network gt_engine to the vehicle
-    vehicle.append_component(turbojet)      
-    
-    
-    # ------------------------------------------------------------------
-    #   Vehicle Definition Complete
-    # ------------------------------------------------------------------
-
-    return vehicle
-
-
-# ----------------------------------------------------------------------
-#   Define the Configurations
-# ---------------------------------------------------------------------
-
-def configs_setup(vehicle):
-    
-    # ------------------------------------------------------------------
-    #   Initialize Configurations
-    # ------------------------------------------------------------------
-    
-    configs = SUAVE.Components.Configs.Config.Container()
-    
-    base_config = SUAVE.Components.Configs.Config(vehicle)
-    base_config.tag = 'base'
-    configs.append(base_config)
-    
-    # ------------------------------------------------------------------
-    #   Cruise Configuration
-    # ------------------------------------------------------------------
-    
-    config = SUAVE.Components.Configs.Config(base_config)
-    config.tag = 'cruise'
-    
-    configs.append(config)
-    
-    
-    # ------------------------------------------------------------------
-    #   Takeoff Configuration
-    # ------------------------------------------------------------------
-    
-    config = SUAVE.Components.Configs.Config(base_config)
-    config.tag = 'takeoff'
-    
-    config.wings['main_wing'].flaps.angle = 0. * Units.deg
-    config.wings['main_wing'].slats.angle = 0. * Units.deg
-    
-    config.V2_VS_ratio = 1.21
-    config.maximum_lift_coefficient = 2.
-    
-    configs.append(config)
-    
-    
-    # ------------------------------------------------------------------
-    #   Landing Configuration
-    # ------------------------------------------------------------------
-
-    config = SUAVE.Components.Configs.Config(base_config)
-    config.tag = 'landing'
-    
-    config.wings['main_wing'].flaps_angle = 0. * Units.deg
-    config.wings['main_wing'].slats_angle = 0. * Units.deg
-
-    config.Vref_VS_ratio = 1.23
-    config.maximum_lift_coefficient = 2.
-    
-    configs.append(config)
-    
-    # done!
-    return configs
-
-# ----------------------------------------------------------------------
-#   Sizing for the Vehicle Configs
-# ----------------------------------------------------------------------
 
 # ----------------------------------------------------------------------
 #   Plot Mission
